refactor: Logo 컴포넌트 및 logo.css추가
